+++ client/package.json
@@ -12,6 +12,8 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "axios": "^0.20.0",
+    "lodash": "^4.17.20",
     "markdown-to-jsx": "^6.11.4",
     "node-sass": "^4.14.1",
     "prop-types": "^15.7.2",
