+++ client/src/components/holders/Main.js
@@ -4,7 +4,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
 import Divider from "@material-ui/core/Divider";
-import Markdown from "./Markdown";
+import Markdown from "../../pages/blog/Markdown";
 import ClapButton from "react-clap-button";
 import { Button } from "@material-ui/core";
