+++ client/src/data/hooks/index.js
@@ -1,0 +1,4 @@
+import useFetchPosts  from './posts'
+
+
+export default useFetchPosts
